refactor: 기존 컴포넌트 및 스토리 alias 로 경로 변경 반영 (#21)
