+++ packages/fds/src/components/Button/Button.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { Slot } from '@radix-ui/react-slot'
 import { cva, type VariantProps } from 'class-variance-authority'
 
-import { cn } from '../../lib/utils'
+import { cn } from '@/lib/utils'
 
 const buttonVariants = cva(
   'w-full flex flex-col items-center justify-center whitespace-nowrap disabled:pointer-events-none cursor-pointer',
