+++ packages/fds/src/components/Form/Form.tsx
@@ -12,7 +12,7 @@ import {
 } from 'react-hook-form'
 
 import { cn } from '@/lib/utils'
-import { Label } from '@/components/Label/Label'
+import { Label } from '@/components'
 
 const Form = FormProvider
