+++ packages/fds/src/components/RadioGroup/RadioGroup.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import * as RadioGroupPrimitive from '@radix-ui/react-radio-group'
 import { CircleIcon } from 'lucide-react'
 
-import { cn } from '../../lib/utils'
+import { cn } from '@/lib/utils'
 
 function RadioGroup({
   className,
